+++ app/services/stern/repair.rb
@@ -89,6 +89,7 @@ module Stern
 
       Entry.delete_all
       EntryPair.delete_all
+      OperationAttempt.delete_all
       Operation.delete_all
       ScheduledOperation.delete_all
     end